==> 01_ReformattingData.ipynb <==
import pandas as pd
import numpy as np

from math import pi
import time


# Read csvs

ings = pd.read_csv('data/og/ingredients.csv', encoding = "ISO-8859-1", index_col=0)
#drop duplicates that have same id and ingredient (they wouldn't be listed twice)
ings = ings.drop_duplicates(subset = ['ingredient', 'id'])

prods = pd.read_csv('data/og/products.csv', encoding = "ISO-8859-1", index_col=0)
#drop the same products
prods = prods.drop_duplicates(subset = 'product')

print('Number of products: ', len(prods))
print('Number of ingredients: ', len(ings))
print('Scraping took 3 days 23:31:45 at 17.27 seconds per product')

##CHECKPOINT
#prods.loc[prods['id'] == 14450]
#ings.loc[ings['id'] == 14450]

# Products

## Reorder columns

# I want the columns in a more natural order
prods = prods[['id', 'product', 'brand', 'price', 'rating', 'ratingsCount', 'categories', 'notes']]

# Remove all spaces for easier cleaning
prods['categories'] = prods['categories'].str.replace(' ', '')
prods['notes'] = prods['notes'].str.replace(' ', '')
prods['price'] = prods['price'].replace({'\$':'', ' ': ''}, regex = True)
prods['price'] = prods['price'].replace({'Nan':'NaN'}, regex = True)
prods['price'] = prods['price'].str.strip()
prods['price'] = prods['price'].astype(str).astype(float)


##Culling of products that aren't Face or Body Skincare
#prods[~prods['categories'].str.contains("Skincare", case=True, flags=0, regex=True)]
prods = prods[prods['categories'].str.contains("FaceSkincare*", na=False)]

prods.head(5)

##CHECKPOINT
prods.loc[prods['id'] == 14450]  #we've dropped the product now because it isn't faceSkincare
#ings.loc[ings['id'] == 14450]

products_keep = prods['id'].tolist()
print('There are', len(prods), 'products') ##now have list of id's to keep

### Convert 'categories' and 'notes' into dummy data

#create categories dataframe that we can use to analyze and join to product dataframe
categories = prods['categories'].str.get_dummies(sep=',')
categories = categories.add_prefix('cat_')
#Drop skincare column as we don't need it since all will be one
categories = categories.drop(['cat_FaceSkincare'], axis =1)
#Drop columns if it has less than 100 products
categories.drop([col for col, val in categories.sum().iteritems() if val < 100], axis=1, inplace=True)

#create notes dataframe that we can use to analyze and join to product dataframe
notes_all = prods['notes'].str.get_dummies(sep=',')
notes_all = notes_all.add_prefix('note_')
notes_keep = notes_all.columns.isin(['note_SulfateFree',
                  'note_ParabenFree',
                  'note_AlcoholFree',
                  'note_Anti-Aging',
                  'note_PromotesWoundHealing',
                  'note_Brightening',
                  'note_Acne-Fighting',
                  'note_UVProtection',
                  'note_ContainsAlcohol',
                  'note_ContainsParaben',
                  'note_ContainsSulfate'])

notes = notes_all.loc[:,notes_keep]

#join to create product dataframe
prods = pd.concat([prods, categories, notes] , axis=1)
# Drop repeated columns 
prods = prods.drop(['categories', 'notes'], axis =1)

# 225 down to 44 columns
prods.head()

Get list of remaining product id's to cull ingredients

### Use id's of both ingredients and products to ensure we have the same products

Create list of unique product id's to cull ingredient id's

id_list = prods['id'].tolist()
print('There are', len(id_list), 'unique products ids')


ings = ings[ings['id'].isin(id_list)]
ing_ids_list = ings['id'].unique()
print('There are', len(ing_ids_list), 'unique ingredient ids')

prods = prods[prods['id'].isin(ing_ids_list)]
prods_ids_list = prods['id'].unique()
print('There are', len(prods_ids_list), 'unique products')

#check unique id differences.
diff = set(prods_ids_list) - set(ing_ids_list)
diff

# Create csv of wide data set
prods.to_csv('data/prods.csv')

### Categories summary

# Convert from wide to long and rename default columns
categories_long = categories.stack().reset_index().rename(columns = {'level_0': 'id', 'level_1': 'category', 0: 'count'})
# Drop all rows where the count == 0 (don't need to count these)
categories_long = categories_long[categories_long['count'] != 0]
# Drp the 'cat_' prefix we used to create dummy data
categories_long['category'] = categories_long['category'].replace({'cat_': ''}, regex = True)
categories_count = pd.DataFrame(categories_long['category'].value_counts())
categories_count['%'] = categories_count['category']/categories_count['category'].sum()*100
categories_count.head()

print('There are', len(categories_count), 'categories')

### Notes summary

# Convert from wide to long and rename default columns
notes_long = notes.stack().reset_index().rename(columns = {'level_0': 'id', 'level_1': 'note', 0: 'count'})
# Drop all rows where the count == 0 (don't need to count these)
notes_long = notes_long[notes_long['count'] != 0]
# Drp the 'note_' prefix we used to create dummy data
notes_long['note'] = notes_long['note'].replace({'note_': ''}, regex = True)
notes_count = pd.DataFrame(notes_long['note'].value_counts())
notes_count['%'] = notes_count['note']/notes_count['note'].sum()*100
notes_count

#notes_count.to_csv('data/notes_count.csv')
print('There are', len(notes_count), 'notes.')

### Ratings summary

# Most products don't have a rating... might drop this further down the line
prods['rating'].value_counts().head()

# Ingredients

Get the columns in a more natural order and strip()

# I want the columns in a more natural order
ings = ings[['id', 'ingredient', 'ewg', 'cir', 'function', 'notes']]
ings['ingredient'] = ings['ingredient'].str.replace('\r\r\n()', '').str.strip('()')
ings['notes'] = ings['notes'].str.replace('\r\r\n\r\r\n', ', ').str.strip()
ings.head()

##CHECKPOINT
#prods.loc[prods['id'] == 14450]  
ings.loc[ings['id'] == 14450] # ing at this id shouldn't exist

prod_ing = ings[['id', 'ingredient']]

#prod_ing.to_csv('data/prod_ing.csv')

#### Summary

ings['ingredient'].value_counts()

print('There are', len(ings), 'ingredients in the database with', ings['ingredient'].nunique(), 'unique ingredients')

### Drop duplicates
Want the ingredients dataframe to just have unique ingredients

ings = ings.drop_duplicates(subset = 'ingredient')
print(len(ings))

#ings.loc[ings['id'] == 3477].head(3)

### Function

Look at just function

ings['function'] = ings['function'].str.replace('- ', '').str.strip('.').str.strip().str.replace(', ', ',')
function = ings['function'].str.get_dummies(sep=',')
function = function.add_prefix('func_')

#down from 216 to 28 when removing less than 100
function.drop([col for col, val in function.sum().iteritems() if val < 100], axis=1, inplace=True)
function.head()

Drop functions with a count less than 200

### Function clean 

#join to create product dataframe
ings = pd.concat([ings, function] , axis=1)
# Drop repeated columns 
ings = ings.drop(['function'], axis =1)
ings.head(3)

### Notes

ings['notes'] = ings['notes'].str.replace(', ', ',').str.strip()
notes = ings['notes'].str.get_dummies(sep=',')
notes = notes.add_prefix('notes_')

notes.head()

#join to create ingredient dataframe
ings = pd.concat([ings, notes] , axis=1)
# Drop repeated columns 
ings = ings.drop(['notes'], axis =1)
ings = ings.drop(['id'], axis =1)
ings.head(3)

ings.to_csv('data/ings.csv')

### Notes summary

# Convert from wide to long and rename default columns
notes_long = notes.stack().reset_index().rename(columns = {'level_0': 'id', 'level_1': 'notes', 0: 'count'})

# Drop all rows where the count == 0 (don't need to count these)
notes_long = notes_long[notes_long['count'] != 0]
notes_long['notes'] = notes_long['notes'].replace({'notes_': ''}, regex = True)
notes_count = pd.DataFrame(notes_long['notes'].value_counts())
notes_count['%'] = notes_count['notes']/notes_count['notes'].sum()*100
notes_count


### Functions summary


# Convert from wide to long and rename default columns
function_long = function.stack().reset_index().rename(columns = {'level_0': 'id', 'level_1': 'function', 0: 'count'})

# Drop all rows where the count == 0 (don't need to count these)
function_long = function_long[function_long['count'] != 0]

# Drp the 'func_' prefix we used to create dummy data
function_long['function'] = function_long['function'].replace({'func_': ''}, regex = True)
function_count = pd.DataFrame(function_long['function'].value_counts())
function_count['%'] = function_count['function']/function_count['function'].sum()*100
function_count

# Product + Ingredients

Need to add a column to retain the order

#prod_ing.loc[prod_ing['id'] == idx]

prod_ing = prod_ing.reset_index(drop=True)

prod_ing['diff'] = prod_ing['id'] - prod_ing['id'].shift(-1)
prod_ing.tail(3)

##CHECKPOINT
prod_ing.loc[prod_ing['id'] == 14450]  
#ings.loc[ings['id'] == 14450] # ing at this id shouldn't exist

%%time
prod_ing['order'] = 0
prod_id = 0
count = 1
for i in range(len(prod_ing)):
    diff = prod_ing.loc[i, ['diff']]
    prod_ing.loc[i, ['order']] = count
    
    if diff.any() == -1:
        count = 1
    else:
        count += 1

#prod_ing.to_csv('data/prod_ing.csv')

df2 = df
df2 = df2.drop(['product', 'ing#List', 'brand', 'price', 'rating', 'ratingsCount', 'ingCount', 'id2'], axis =1)

df2['ingList'] = [str(i)[1:-1] for i in df2['ingList']]

##create dummy df of ingredients
ings_all = df2['ingList'].str.get_dummies(sep=',')
ings_all = ings_all.add_prefix('ing_')

ings_all.to_csv('data/prodIngs_Wide.csv')

len(prod_ing)

prod_ing.head()
