==> 05_VizResults.ipynb <==
import pandas as pd
import numpy as np

from math import pi
import time


df = pd.read_csv('data/VizResults.csv')
df_key = pd.read_csv('data/VizResults_key.csv')
df.head()

df.pivot('country','listing_neighborhood','count').plot(stacked=True, kind='bar')


Q = 'Q44'
df[Q].value_counts()

df['Q46']
